Fix that pesky stack bug
I was giving v8 the top of the stack instead of the bottom of the stack
so it couldn't figure out where to stop. This change fixes that. I
tested this using an infinite recussion function with really small
stacks (8kb) and large ones (512k) and I'm getting reasonable results.

Now the hard part is picking the right stack size. I'm going to go with
64k as it's the default stack node picked for libeio threads.