CORO_PTHREAD compatibility
If for some reason UCONTEXT and SJLJ don't work, you can change the
CORO_WHATEVER definition in src/platform.mk to CORO_PTHREAD and it will
use pthreads and pthread_cond_wait to simulate fibers which should be
pretty resilient against v8 changes. This isn't being used for now but
it only needed a few hooks to get this compatibility in.